requirements.txt
sentence-transformers==2.5.1
annoy==1.17.3
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
rich==13.7.0
typer==0.9.0
python-dotenv==1.0.1
unidiff==0.7.5